+++ scripts/data/get_data.sh
@@ -1,5 +1,5 @@
 #!/bin/bash
 git clone https://github.com/RaccoonResearch/simdata
-mv simdata/filtered datasets
+mv simdata/datasets datasets
 mv simdata/examples examples
 rm -rf simdata
