+++ simian/batch.py
@@ -10,8 +10,8 @@ from questionary import Choice, select
 from rich.console import Console
 from typing import Any, Dict, List, Optional
 
-from simian.prompts import generate_gemini, setup_gemini, CAMERA_PROMPT, OBJECTS_JSON_PROMPT, OBJECTS_PROMPT
-from .server import initialize_chroma_db, query_collection, process_in_batches
+from simian.prompts import generate_gemini, setup_gemini, parse_gemini_json, CAMERA_PROMPT, OBJECTS_JSON_PROMPT, OBJECTS_PROMPT
+from .server import initialize_chroma_db, query_collection
 from .combiner import calculate_transformed_positions
 
 console = Console()
@@ -214,42 +214,7 @@ def parse_args(args_list = None) -> argparse.Namespace:
     return args
 
 
-def parse_gemini_json(raw_output: str) -> Optional[dict]:
-    """
-    Parse the JSON output from the Gemini API.
 
-    Args:
-        raw_output (str): The raw output from the Gemini API.
-
-    Returns:
-        Optional[dict]: The parsed JSON output as a dictionary, or None if an error occurs.
-    """
-
-    try:
-        if "```json" in raw_output:
-            json_start = raw_output.index("```json") + 7
-            json_end = raw_output.rindex("```")
-            json_content = raw_output[json_start:json_end].strip()
-        else:
-            json_content = raw_output.strip()
-
-        # Remove any leading or trailing commas
-        json_content = json_content.strip(',')
-
-        # If the content starts with a key (e.g., "objects":), wrap it in curly braces
-        if json_content.strip().startswith('"') and ':' in json_content:
-            json_content = "{" + json_content + "}"
-
-        # Parse the JSON
-        parsed_json = json.loads(json_content)
-        
-        print("Successfully parsed JSON:")
-        print(json.dumps(parsed_json, indent=2))
-        
-        return parsed_json
-    except json.JSONDecodeError as e:
-        print(f"Error parsing JSON: {str(e)}")
-        return None
 
 
 def prompt_based_rendering():
